gfx: add gfx_get_glyph

+++ code/gfx/gfx_text.hpp
@@ -15,7 +15,7 @@ struct GFX_Glyph {
 };
 
 must_use global GFX_Font
-gfx_make_font(GFX_Image image, u32 char_width, u32 char_height);
+gfx_make_font(GFX_Image image, u16 char_width, u16 char_height);
 
 must_use global GFX_Glyph
-gfx_get_glyph(GFX_Font *font, u8 ch, f32 height_px);
+gfx_get_glyph(GFX_Font *font, u8 ch, u16 height_px);
